Feedback checklist
Where your_namespace is used can this be changed to project_namespace (x)

Can you look at the incorporation of the queue mechanism into this - there has to be some element of how this will affect the quotas and use of resources?

Could this be move out of training and into another directory under the gpuservice (data_transfer or similar)? (x)

Is there a security risk here if any access tokens or similar as stored in the PVC inside a project? (x)