Open OnDemand JupyterLab and RStudio container apps no longer raise 'Proxy Error'
Bug fixes have been applied to JupyterLab and RStudio container apps so that the 'Connect' button is not rendered to users until the services running in these containers are ready to receive connections. This means that the 'Proxy Error' that formerly arose when users clicked on the 'Connect' button when the services were not ready, no longer occurs.

Remove the relevant 'Troubleshooting' from `docs/safe-haven-services/open-ondemand/` `getting-started.md`, `apps/jupyter-app.md` and `appsr/studio-app.md`.